requirements.txt
llama-index==0.12.28
matplotlib==3.10.1
streamlit==1.44.1
pandas==2.2.3
langchain==0.3.22
ollama==0.4.7
langchain-ollama==0.3.0
langgraph==0.3.25
transformers==4.51.2
langchain-core==0.3.51
duckduckgo-search==8.0.1
langchain-community==0.3.20
playwright==1.51.0
lxml==5.3.2
langchain-tavily==0.1.6
python-dotenv==1.1.0